Refactor core engine, add World & .NET cache
This commit again by copilot because idk what to write i has forget all what was doing

Introduce a major core refactor: add World and entity/component primitives (Entity, IComponent, Transform, SpatialGrid, World), physics/collision support (new Collision plus PhysicsBody/Forces/PhysicsSystem updates), rendering/input improvements and many extension utilities (Array/Span/String/Vector3/Disposable). Also add ConsoleUtils, ScreenShakeSystem and Test.cs, and update demo, Program.cs and the project file. In addition, add numerous .dotnet user cache/sentinel and TelemetryStorageService files (local .NET environment and telemetry artifacts). These changes prepare the engine for component-based world management and improved physics/rendering integration.

+++ Core/SystemsUpdater.cs
@@ -7,11 +7,11 @@ namespace SomeSimpleConsoleGame.Core
     {
         bool IsReady() => true;
 
-        void Update(double deltaTime);
+        void Update(float deltaTime);
     }
     public sealed class SystemsUpdater : IDisposable
     {
-        private static readonly double TicksToSeconds = 1.0 / Stopwatch.Frequency;
+        private static readonly float TicksToSeconds = 1f / Stopwatch.Frequency;
         private static readonly Comparer<(IUpdateSystem system, byte priority, long lastCallTicks)> PriorityComparer =
             Comparer<(IUpdateSystem, byte priority, long)>.Create(static (a, b) => b.priority.CompareTo(a.priority));
